requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.14.54
botocore==1.17.54
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.43.0
django-cors-headers==3.4.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
django-storages==1.10
django-webpack-loader==0.7.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jimmy-util @ git+https://github.com/seungjinhan/python_jimmy_util.git@0f7c7c407bbb7214469406d82dfc83a9a308ac8a
jmespath==0.10.0
lazy-object-proxy==1.4.3
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
